requirements_v2.txt
colorama>=0.4.6
requests>=2.28.0
paramiko>=3.0.0
dnspython>=2.4.0
cryptography>=41.0.0
Pillow>=10.0.0
